+++ lib/src/base/base.dart
@@ -1,3 +1,4 @@
 export 'bloc/bloc.dart';
 export 'event/event.dart';
 export 'state/state.dart';
+export 'model/model.dart';