+++ lib/src/base/bloc/base_load_bloc.dart
@@ -1,5 +1,5 @@
 import 'package:bloc/bloc.dart';
-import 'package:iyam_core/iyam_core.dart';
+import 'package:eitherx/eitherx.dart';
 import '../event/load_event.dart';
 import '../state/base_state.dart';
 import '../state/load_state.dart';
@@ -10,17 +10,16 @@ abstract class BaseLoadBloc<T> extends Bloc<LoadEvent, BaseState> {
     on<RefreshRequested>(_onRefresh);
   }
 
-  Future<RepositoryResult<T>> fetch();
+  Future<Either<String, T>> fetch();
 
   Future<void> _onLoad(LoadRequested event, Emitter<BaseState> emit) async {
     emit(LoadingState());
 
     final result = await fetch();
-    if (result is RepoSuccess<T>) {
-      emit(SuccessState<T>(result.data));
-    } else if (result is RepoFailure<T>) {
-      emit(ErrorState(result.message));
-    }
+    result.fold(
+      (error) => emit(ErrorState(error)),
+      (data) => emit(SuccessState<T>(data)),
+    );
   }
 
   Future<void> _onRefresh(
@@ -28,8 +27,6 @@ abstract class BaseLoadBloc<T> extends Bloc<LoadEvent, BaseState> {
     Emitter<BaseState> emit,
   ) async {
     final result = await fetch();
-    if (result is RepoSuccess<T>) {
-      emit(SuccessState<T>(result.data));
-    }
+    result.fold((error) {}, (data) => emit(SuccessState<T>(data)));
   }
 }